# 案例4：Redis常见数据结构实战-RedisTemplate实战字符串数据结构和单个实体对象数据结构/单个实体对象/fastPower/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: 系统 V1.0 API接口文档
  version: 1.0.0
  title: 接口文档
basePath: /fast
schemes:
  - http
securityDefinitions:
  api_key:
    type: apiKey
    name: token
    in: header
paths:
...
  /sys/login:
    post:
      tags:
        - 后台用户中心
      summary: 后台用户登录获取token令牌
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理员对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/LoginForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Login'
...
definitions:
...
  Login:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      token:
        description: token
        type: string
      expire:
        description: 过期时长
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  LoginForm:
    type: object
    properties:
      username:
        description: 用户名
        type: string
      password:
        description: 密码
        type: string
      captcha:
        description: 验证码
        type: string
      uuid:
        description: UUID
        type: string
...